all: split process package into separate process and platform packages
