+++ pkg/platform/doc.go
@@ -1,0 +1,2 @@
+// Package platform provides platform-specific definitions and functionality.
+package platform
