+++ pkg/platform/executable_name_test.go
@@ -1,4 +1,4 @@
-package process
+package platform
 
 import (
 	"testing"
