+++ pkg/platform/find_command_test.go
@@ -1,3 +1,3 @@
-package process
+package platform
 
 // TODO: Implement.
